+++ src/coordinates.cpp
@@ -24,18 +24,18 @@ double lat2latp(double lat) { return rad2deg(log(tan(deg2rad(clamp(lat,85.06)+90
 double latp2lat(double latp) { return rad2deg(atan(exp(deg2rad(latp)))*2.0)-90.0; }
 
 // Tile conversions
-double lon2tilexf(double lon, uint z) { return scalbn((lon+180.0) * (1/360.0), (int)z); }
-double latp2tileyf(double latp, uint z) { return scalbn((180.0-latp) * (1/360.0), (int)z); }
-double lat2tileyf(double lat, uint z) { return latp2tileyf(lat2latp(lat), z); }
-uint lon2tilex(double lon, uint z) { return lon2tilexf(lon, z); }
-uint latp2tiley(double latp, uint z) { return latp2tileyf(latp, z); }
-uint lat2tiley(double lat, uint z) { return lat2tileyf(lat, z); }
-double tilex2lon(uint x, uint z) { return scalbn(x, -(int)z) * 360.0 - 180.0; }
-double tiley2latp(uint y, uint z) { return 180.0 - scalbn(y, -(int)z) * 360.0; }
-double tiley2lat(uint y, uint z) { return latp2lat(tiley2latp(y, z)); }
+double lon2tilexf(double lon, uint8_t z) { return scalbn((lon+180.0) * (1/360.0), (int)z); }
+double latp2tileyf(double latp, uint8_t z) { return scalbn((180.0-latp) * (1/360.0), (int)z); }
+double lat2tileyf(double lat, uint8_t z) { return latp2tileyf(lat2latp(lat), z); }
+uint32_t lon2tilex(double lon, uint8_t z) { return lon2tilexf(lon, z); }
+uint32_t latp2tiley(double latp, uint8_t z) { return latp2tileyf(latp, z); }
+uint32_t lat2tiley(double lat, uint8_t z) { return lat2tileyf(lat, z); }
+double tilex2lon(uint32_t x, uint8_t z) { return scalbn(x, -(int)z) * 360.0 - 180.0; }
+double tiley2latp(uint32_t y, uint8_t z) { return 180.0 - scalbn(y, -(int)z) * 360.0; }
+double tiley2lat(uint32_t y, uint8_t z) { return latp2lat(tiley2latp(y, z)); }
 
 // Get a tile index
-TileCoordinates latpLon2index(LatpLon ll, uint baseZoom) {
+TileCoordinates latpLon2index(LatpLon ll, uint8_t baseZoom) {
 	return TileCoordinates(lon2tilex(ll.lon /10000000.0, baseZoom),
 	       latp2tiley(ll.latp/10000000.0, baseZoom));
 }
